feat: Christmas discount policy can calculate discount amount
