+++ src/main/kotlin/christmas/domain/Price.kt
@@ -6,6 +6,7 @@ data class Price(val value: Double) {
     }
 
     infix fun times(operand: Int) = Price(value * operand)
+    infix fun plus(other: Price) = Price(value + other.value)
 
     companion object {
         fun from(value: Int) = Price(value.toDouble())
