+++ src/main/kotlin/christmas/domain/event/discount/DiscountPolicy.kt
@@ -5,5 +5,5 @@ import christmas.domain.order.Order
 
 interface DiscountPolicy {
     fun isEligibleForDiscount(order: Order): Boolean
-    fun apply(order: Order): Price
+    fun getDiscountAmount(order: Order): Price
 }
